service/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
apscheduler==3.10.4
requests==2.31.0
python-dotenv==1.0.1
pylint==3.1.0